# PHOENIX-6537 Fix random maven-related Jenkins build errors