--- a/dev/create-release/phoenix-rm/Dockerfile
+++ b/dev/create-release/phoenix-rm/Dockerfile
@@ -46,7 +46,7 @@ RUN DEBIAN_FRONTEND=noninteractive apt-get -qq -y update \
   && pip install \
     python-dateutil==2.8.1
 # Install Apache Yetus
-ENV YETUS_VERSION 0.12.0
+ENV YETUS_VERSION 0.13.0
 SHELL ["/bin/bash", "-o", "pipefail", "-c"]
 RUN wget -qO- "https://www.apache.org/dyn/mirrors/mirrors.cgi?action=download&filename=/yetus/${YETUS_VERSION}/apache-yetus-${YETUS_VERSION}-bin.tar.gz" | \
         tar xvz -C /opt