--- a/dev/docker/Dockerfile.multibranch
+++ b/dev/docker/Dockerfile.multibranch
@@ -14,7 +14,17 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-FROM apache/yetus:0.12.0
+FROM apache/yetus:0.13.0
 
 RUN groupadd --non-unique -g 910 jenkins || true
 RUN useradd -g 910 -u 910 -m jenkins || true
+#####
+# /run/user directory for gpg
+#####
+RUN mkdir -p /run/user/910/gnupg/ \
+    && chmod -R 700 /run/user/910 \
+    && chown -R 910:910 /run/user/910 \
+    && mkdir -p /var/run/user/910/gnupg \
+    && chown -R 910:910 /var/run/user/910 \
+    && chmod -R 700 /var/run/user/910
+ENV JAVA_HOME /usr/lib/jvm/java-11-openjdk-amd64