--- a/phoenix-core/src/it/java/org/apache/phoenix/end2end/Array3IT.java
+++ b/phoenix-core/src/it/java/org/apache/phoenix/end2end/Array3IT.java
@@ -32,7 +32,9 @@ import java.util.Properties;
 
 import org.apache.phoenix.util.PropertiesUtil;
 import org.junit.Test;
+import org.junit.experimental.categories.Category;
 
+@Category(ParallelStatsDisabledTest.class)
 public class Array3IT extends ArrayIT {
 
     @Test