… phoenix-core/src/it/java/org/apache/phoenix/end2end/BaseUniqueNamesOwnClusterIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/CountDistinctApproximateHyperLogLogIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/CountDistinctCompressionIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/CreateTableIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/CsvBulkLoadToolIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/CursorWithRowValueConstructorIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/CustomEntityDataIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/DateArithmeticIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/DateTimeIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/DecodeFunctionIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/DefaultColumnValueIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/DeleteIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/DerivedTableIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/DescOrderKeysIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/DisableLocalIndexIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/DistinctCountIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/DistinctPrefixFilterIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/DropIndexedColsIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/DropSchemaIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/DropTableIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/DropTableWithViewsIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/DynamicColumnIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/DynamicColumnWildcardIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/DynamicFamilyIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/DynamicUpsertIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/EmptyColumnIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/EncodeFunctionIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/EvaluationOfORIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/ExecuteStatementsIT.java
… phoenix-core/src/it/java/org/apache/phoenix/end2end/ExpFunctionEnd2EndIT.java