@stoty 2021-09-09
:broken_heart: **-1 overall**






| Vote | Subsystem | Runtime | Comment |
|:----:|----------:|--------:|:--------|
| +0 :ok: |  reexec  |   7m 15s |  Docker mode activated.  |
||| _ Prechecks _ |
| +1 :green_heart: |  dupname  |   0m  0s |  No case conflicting files found.  |
| +1 :green_heart: |  @author  |   0m  0s |  The patch does not contain any @author tags.  |
| -1 :x: |  test4tests  |   0m  0s |  The patch doesn't appear to include any new or modified tests. Please justify why no new tests are needed for this patch. Also please list what manual steps were performed to verify this patch.  |
||| _ master Compile Tests _ |
| +1 :green_heart: |  mvninstall  |  27m 12s |  master passed  |
| +0 |  hbaserecompile  |  30m  5s |  HBase recompiled.  |
| +1 :green_heart: |  compile  |   1m 53s |  master passed  |
| -1 :x: |  javadoc  |   0m 47s |  root in master failed.  |
||| _ Patch Compile Tests _ |
| +1 :green_heart: |  mvninstall  |  18m 34s |  the patch passed  |
| +0 |  hbaserecompile  |  22m 18s |  HBase recompiled.  |
| +1 :green_heart: |  compile  |   1m 50s |  the patch passed  |
| +1 :green_heart: |  javac  |   1m 50s |  the patch passed  |
| +1 :green_heart: |  hadolint  |   0m  4s |  There were no new hadolint issues.  |
| +1 :green_heart: |  shellcheck  |   0m  0s |  There were no new shellcheck issues.  |
| +1 :green_heart: |  shelldocs  |   0m  2s |  There were no new shelldocs issues.  |
| +1 :green_heart: |  whitespace  |   0m  0s |  The patch has no whitespace issues.  |
| +1 :green_heart: |  xml  |   0m  2s |  The patch has no ill-formed XML file.  |
| -1 :x: |  javadoc  |   0m 45s |  root in the patch failed.  |
||| _ Other Tests _ |
| -1 :x: |  unit  |   2m 12s |  root in the patch failed.  |
| +1 :green_heart: |  asflicense  |   0m 10s |  The patch does not generate ASF License warnings.  |
|  |   |  67m  5s |   |


| Reason | Tests |
|-------:|:------|
| Failed junit tests | phoenix.hbase.index.write.recovery.TestPerRegionIndexWriteCache |


| Subsystem | Report/Notes |
|----------:|:-------------|
| Docker | ClientAPI=1.41 ServerAPI=1.41 base: https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1295/1/artifact/yetus-general-check/output/Dockerfile |
| GITHUB PR | https://github.com/apache/phoenix/pull/1295 |
| Optional Tests | dupname asflicense hadolint shellcheck shelldocs javac javadoc unit xml compile |
| uname | Linux d9b0295216da 4.15.0-153-generic #160-Ubuntu SMP Thu Jul 29 06:54:29 UTC 2021 x86_64 x86_64 x86_64 GNU/Linux |
| Build tool | maven |
| Personality | dev/phoenix-personality.sh |
| git revision | master / 19b22604e |
| Default Java | Ubuntu-11.0.9.1+1-Ubuntu-0ubuntu1.20.04 |
| javadoc | https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1295/1/artifact/yetus-general-check/output/branch-javadoc-root.txt |
| javadoc | https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1295/1/artifact/yetus-general-check/output/patch-javadoc-root.txt |
| unit | https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1295/1/artifact/yetus-general-check/output/patch-unit-root.txt |
|  Test Results | https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1295/1/testReport/ |
| Max. process+thread count | 473 (vs. ulimit of 30000) |
| modules | C: . U: . |
| Console output | https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1295/1/console |
| versions | git=2.25.1 maven=3.6.3 shellcheck=0.7.1 hadolint=1.18.0-0-g76eee5c |
| Powered by | Apache Yetus 0.12.0 https://yetus.apache.org |


This message was automatically generated.