@stoty 2021-09-13
:broken_heart: **-1 overall**






| Vote | Subsystem | Runtime | Comment |
|:----:|----------:|--------:|:--------|
| +0 :ok: |  reexec  |   1m  7s |  Docker mode activated.  |
||| _ Prechecks _ |
| +1 :green_heart: |  dupname  |   0m  0s |  No case conflicting files found.  |
| +1 :green_heart: |  @author  |   0m  0s |  The patch does not contain any @author tags.  |
| -1 :x: |  test4tests  |   0m  0s |  The patch doesn't appear to include any new or modified tests. Please justify why no new tests are needed for this patch. Also please list what manual steps were performed to verify this patch.  |
||| _ master Compile Tests _ |
| +1 :green_heart: |  mvninstall  |  28m 20s |  master passed  |
| +0 |  hbaserecompile  |  37m 53s |  HBase recompiled.  |
| +1 :green_heart: |  compile  |   2m  6s |  master passed  |
| -1 :x: |  javadoc  |   1m  2s |  root in master failed.  |
||| _ Patch Compile Tests _ |
| +1 :green_heart: |  mvninstall  |  18m 37s |  the patch passed  |
| +0 |  hbaserecompile  |  27m 36s |  HBase recompiled.  |
| +1 :green_heart: |  compile  |   1m 56s |  the patch passed  |
| +1 :green_heart: |  javac  |   1m 56s |  the patch passed  |
| +1 :green_heart: |  hadolint  |   0m  3s |  There were no new hadolint issues.  |
| -1 :x: |  shellcheck  |   0m  1s |  The patch generated 1 new + 0 unchanged - 0 fixed = 1 total (was 0)  |
| +1 :green_heart: |  shelldocs  |   0m  4s |  There were no new shelldocs issues.  |
| +1 :green_heart: |  whitespace  |   0m  0s |  The patch has no whitespace issues.  |
| +1 :green_heart: |  xml  |   0m  2s |  The patch has no ill-formed XML file.  |
| -1 :x: |  javadoc  |   0m 50s |  root in the patch failed.  |
||| _ Other Tests _ |
| -1 :x: |  unit  |   6m 54s |  root in the patch failed.  |
| -1 :x: |  asflicense  |   0m 24s |  The patch generated 613 ASF License warnings.  |
|  |   |  79m 57s |   |


| Reason | Tests |
|-------:|:------|
| Failed junit tests | phoenix.end2end.RegexBulkLoadToolIT |
|   | phoenix.end2end.index.GlobalMutableNonTxIndexWithLazyPostBatchWriteIT |
|   | phoenix.end2end.StoreNullsPropIT |
|   | phoenix.end2end.NumericArithmeticIT |
|   | phoenix.end2end.index.txn.RollbackIT |
|   | phoenix.end2end.index.PartialIndexRebuilderIT |
|   | phoenix.end2end.DistinctPrefixFilterIT |
|   | phoenix.end2end.TenantIdTypeIT |
|   | phoenix.end2end.SkipScanAfterManualSplitIT |
|   | phoenix.end2end.IndexToolTimeRangeIT |
|   | phoenix.tx.ParameterizedTransactionIT |
|   | phoenix.end2end.AuditLoggingIT |
|   | phoenix.end2end.ShowCreateTableIT |
|   | phoenix.end2end.IndexRebuildTaskIT |
|   | phoenix.end2end.TableSnapshotReadsMapReduceIT |
|   | phoenix.tx.FlappingTransactionIT |
|   | phoenix.iterate.RoundRobinResultIteratorWithStatsIT |
|   | phoenix.end2end.PermissionsCacheIT |
|   | phoenix.end2end.TenantSpecificViewIndexSaltedIT |
|   | phoenix.end2end.SpillableGroupByIT |
|   | phoenix.end2end.RenewLeaseIT |
|   | phoenix.end2end.RegexpReplaceFunctionIT |
|   | phoenix.end2end.ViewConcurrencyAndFailureIT |
|   | phoenix.end2end.ToNumberFunctionIT |
|   | phoenix.iterate.ScannerLeaseRenewalIT |
|   | phoenix.rpc.PhoenixClientRpcIT |
|   | phoenix.end2end.CreateSchemaIT |
|   | phoenix.hbase.index.FailForUnsupportedHBaseVersionsIT |
|   | phoenix.monitoring.PhoenixLoggingMetricsIT |
|   | phoenix.query.ConnectionCachingIT |
|   | phoenix.end2end.IndexExtendedIT |
|   | phoenix.end2end.UpsertBigValuesIT |
|   | phoenix.end2end.QueryIT |
|   | phoenix.end2end.salted.SaltedTableUpsertSelectIT |
|   | phoenix.end2end.ReverseFunctionIT |
|   | phoenix.replication.SystemCatalogWALEntryFilterIT |
|   | phoenix.end2end.DynamicFamilyIT |
|   | phoenix.execute.UpsertSelectOverlappingBatchesIT |
|   | phoenix.end2end.QueryMoreIT |
|   | phoenix.end2end.index.ImmutableIndexExtendedIT |
|   | phoenix.end2end.SetPropertyOnNonEncodedTableIT |
|   | phoenix.end2end.ViewTTLIT |
|   | phoenix.end2end.UpdateCacheAcrossDifferentClientsIT |
|   | phoenix.end2end.index.DropMetadataIT |
|   | phoenix.end2end.ProductMetricsIT |
|   | phoenix.end2end.ImmutableTableIT |
|   | phoenix.end2end.OrderByWithServerClientSpoolingDisabledIT |
|   | phoenix.end2end.StatementHintsIT |
|   | phoenix.end2end.PermissionNSDisabledWithCustomAccessControllerIT |
|   | phoenix.end2end.IndexToolForPartialBuildWithNamespaceEnabledIT |
|   | phoenix.end2end.IntArithmeticIT |
|   | phoenix.end2end.ExecuteStatementsIT |
|   | phoenix.end2end.CountDistinctApproximateHyperLogLogIT |
|   | phoenix.end2end.SystemTablesCreationOnConnectionIT |
|   | phoenix.end2end.AbsFunctionEnd2EndIT |
|   | phoenix.end2end.Array2IT |
|   | phoenix.end2end.PhoenixRuntimeIT |
|   | phoenix.end2end.ImmutableTablePropertiesIT |
|   | phoenix.end2end.RowKeyBytesStringFunctionIT |
|   | phoenix.end2end.ColumnEncodedBytesPropIT |
|   | phoenix.rpc.UpdateCacheConnectionLevelPropIT |
|   | phoenix.end2end.ConnectionUtilIT |
|   | phoenix.end2end.BinaryRowKeyIT |
|   | phoenix.end2end.index.MutableIndexFailureIT |
|   | phoenix.monitoring.PhoenixTableLevelMetricsIT |
|   | phoenix.util.IndexScrutinyIT |
|   | phoenix.end2end.SequenceIT |
|   | phoenix.end2end.ContextClassloaderIT |
|   | phoenix.schema.stats.NonTxStatsCollectorIT |
|   | phoenix.end2end.UpsertSelectIT |
|   | phoenix.end2end.MutationStateIT |
|   | phoenix.end2end.ReadIsolationLevelIT |
|   | phoenix.end2end.ColumnProjectionOptimizationIT |
|   | phoenix.end2end.MultiCfQueryExecIT |
|   | phoenix.end2end.SerialIteratorsIT |
|   | phoenix.end2end.Array3IT |
|   | phoenix.end2end.TruncateFunctionIT |
|   | phoenix.end2end.UpsertSelectAutoCommitIT |
|   | phoenix.end2end.SortMergeJoinMoreIT |
|   | phoenix.iterate.PhoenixQueryTimeoutIT |
|   | phoenix.end2end.join.HashJoinLocalIndexIT |
|   | phoenix.end2end.IndexToolForDeleteBeforeRebuildIT |
|   | phoenix.end2end.StoreNullsIT |
|   | phoenix.end2end.LastValuesFunctionIT |
|   | phoenix.end2end.index.GlobalMutableNonTxIndexIT |
|   | phoenix.end2end.ExtendedQueryExecIT |
|   | phoenix.end2end.ExplainPlanWithStatsEnabledIT |
|   | phoenix.end2end.CollationKeyFunctionIT |
|   | phoenix.end2end.TransactionalViewIT |
|   | phoenix.end2end.MapReduceIT |
|   | phoenix.end2end.UpgradeIT |
|   | phoenix.end2end.EncodeFunctionIT |
|   | phoenix.end2end.SequenceBulkAllocationIT |
|   | phoenix.end2end.ArrayFillFunctionIT |
|   | phoenix.end2end.NativeHBaseTypesIT |
|   | phoenix.end2end.index.ViewIndexIT |
|   | phoenix.end2end.NotQueryWithLocalImmutableIndexesIT |
|   | phoenix.rpc.UpdateCacheIT |
|   | phoenix.end2end.index.LocalMutableTxIndexIT |
|   | phoenix.end2end.DateTimeIT |
|   | phoenix.end2end.DistinctCountIT |
|   | phoenix.end2end.DefaultColumnValueIT |
|   | phoenix.end2end.PointInTimeScanQueryIT |
|   | phoenix.end2end.MigrateSystemTablesToSystemNamespaceIT |
|   | phoenix.end2end.DateArithmeticIT |
|   | phoenix.end2end.IndexVerificationOldDesignIT |
|   | phoenix.end2end.index.LocalIndexIT |
|   | phoenix.end2end.index.IndexVerificationResultRepositoryIT |
|   | phoenix.end2end.InQueryIT |
|   | phoenix.end2end.StringToArrayFunctionIT |
|   | phoenix.end2end.salted.SaltedTableVarLengthRowKeyIT |
|   | phoenix.end2end.ArrayConcatFunctionIT |
|   | phoenix.end2end.ArrayPrependFunctionIT |
|   | phoenix.end2end.index.GlobalImmutableNonTxIndexIT |
|   | phoenix.end2end.AutoCommitIT |
|   | phoenix.end2end.UngroupedIT |
|   | phoenix.end2end.LikeExpressionIT |
|   | phoenix.end2end.IndexToolForPartialBuildIT |
|   | phoenix.end2end.EvaluationOfORIT |
|   | phoenix.end2end.CountDistinctCompressionIT |
|   | phoenix.end2end.CursorWithRowValueConstructorIT |
|   | phoenix.end2end.index.IndexCoprocIT |
|   | phoenix.end2end.ConfigurableCacheIT |
|   | phoenix.end2end.GetSetByteBitFunctionEnd2EndIT |
|   | phoenix.coprocessor.StatisticsCollectionRunTrackerIT |
|   | phoenix.query.MaxConcurrentConnectionsIT |
|   | phoenix.end2end.CSVCommonsLoaderIT |
|   | phoenix.end2end.index.MutableIndexExtendedIT |
|   | phoenix.end2end.OrderByWithServerMemoryLimitIT |
|   | phoenix.end2end.NamespaceSchemaMappingIT |
|   | phoenix.end2end.StatsEnabledSplitSystemCatalogIT |
|   | phoenix.end2end.join.WhereOptimizerForJoinFiltersIT |
|   | phoenix.end2end.ClientHashAggregateIT |
|   | phoenix.end2end.CastAndCoerceIT |
|   | phoenix.end2end.index.MutableIndexSplitForwardScanIT |
|   | phoenix.monitoring.PhoenixMetricsIT |
|   | phoenix.end2end.ArrayAppendFunctionIT |
|   | phoenix.end2end.FunkyNamesIT |
|   | phoenix.end2end.OrderByIT |
|   | phoenix.mapreduce.PhoenixServerBuildIndexInputFormatIT |
|   | phoenix.end2end.QueryWithLimitIT |
|   | phoenix.schema.stats.TxStatsCollectorIT |
|   | phoenix.end2end.index.ChildViewsUseParentViewIndexIT |
|   | phoenix.end2end.ParameterizedIndexUpgradeToolIT |
|   | phoenix.end2end.index.GlobalImmutableTxIndexIT |
|   | phoenix.iterate.RoundRobinResultIteratorIT |
|   | phoenix.end2end.InstrFunctionIT |
|   | phoenix.end2end.SystemCatalogRollbackEnabledIT |
|   | phoenix.end2end.TimezoneOffsetFunctionIT |
|   | phoenix.end2end.ConvertTimezoneFunctionIT |
|   | phoenix.end2end.salted.SaltedTableIT |
|   | phoenix.end2end.ParallelIteratorsIT |
|   | phoenix.end2end.IndexToolIT |
|   | phoenix.end2end.AggregateIT |
|   | phoenix.end2end.RebuildIndexConnectionPropsIT |
|   | phoenix.end2end.PermissionNSEnabledWithCustomAccessControllerIT |
|   | phoenix.end2end.UserDefinedFunctionsIT |
|   | phoenix.end2end.PermissionNSDisabledIT |
|   | phoenix.end2end.AutoPartitionViewsIT |
|   | phoenix.end2end.VariableLengthPKIT |
|   | phoenix.end2end.QueryWithOffsetIT |
|   | phoenix.end2end.OperationTimeoutWithReasonIT |
|   | phoenix.end2end.join.SortMergeJoinNoIndexIT |
|   | phoenix.end2end.IndexBuildTimestampIT |
|   | phoenix.end2end.FlappingLocalIndexIT |
|   | phoenix.end2end.index.ImmutableIndexWithStatsIT |
|   | phoenix.end2end.DropTableIT |
|   | phoenix.end2end.index.IndexWithTableSchemaChangeIT |
|   | phoenix.end2end.SequencePointInTimeIT |
|   | phoenix.end2end.RowValueConstructorOffsetOptionalIT |
|   | phoenix.end2end.index.SingleCellIndexIT |
|   | phoenix.end2end.index.AlterIndexIT |
|   | phoenix.end2end.LogicalTableNameExtendedIT |
|   | phoenix.end2end.UseSchemaIT |
|   | phoenix.end2end.index.DropColumnIT |
|   | phoenix.end2end.CustomEntityDataIT |
|   | phoenix.end2end.ArrayToStringFunctionIT |
|   | phoenix.end2end.join.HashJoinPersistentCacheIT |
|   | phoenix.end2end.ViewMetadataIT |
|   | phoenix.end2end.CsvBulkLoadToolIT |
|   | phoenix.end2end.LastValueFunctionIT |
|   | phoenix.end2end.FlappingAlterTableIT |
|   | phoenix.end2end.SetPropertyOnEncodedTableIT |
|   | phoenix.end2end.PointInTimeQueryIT |
|   | phoenix.end2end.index.LocalMutableNonTxIndexIT |
|   | phoenix.end2end.QueryDatabaseMetaDataIT |
|   | phoenix.end2end.ServerExceptionIT |
|   | phoenix.end2end.DescOrderKeysIT |
|   | phoenix.util.CoprocessorHConnectionTableFactoryIT |
|   | phoenix.end2end.DeleteIT |
|   | phoenix.end2end.ConcurrentUpsertsWithoutIndexedColsIT |
|   | phoenix.end2end.QueryWithTableSampleIT |
|   | phoenix.end2end.PhoenixRowTimestampFunctionIT |
|   | phoenix.end2end.join.SubqueryIT |
|   | phoenix.end2end.index.SaltedIndexIT |
|   | phoenix.end2end.MathTrigFunctionEnd2EndIT |
|   | phoenix.end2end.QueryTimeoutIT |
|   | phoenix.end2end.StddevIT |
|   | phoenix.end2end.index.MutableIndexIT |
|   | phoenix.end2end.RTrimFunctionIT |
|   | phoenix.end2end.TopNIT |
|   | phoenix.end2end.IsNullIT |
|   | phoenix.end2end.RowValueConstructorOffsetIT |
|   | phoenix.end2end.NoLookbackMutableIndexExtendedIT |
|   | phoenix.end2end.StringIT |
|   | phoenix.end2end.DisableLocalIndexIT |
|   | phoenix.end2end.NotQueryWithGlobalImmutableIndexesIT |
|   | phoenix.end2end.ToCharFunctionIT |
|   | phoenix.end2end.LocalIndexSplitMergeIT |
|   | phoenix.end2end.index.IndexAsyncThresholdIT |
|   | phoenix.end2end.SignFunctionEnd2EndIT |
|   | phoenix.end2end.MinMaxAggregateFunctionIT |
|   | phoenix.end2end.join.SortMergeJoinLocalIndexIT |
|   | phoenix.end2end.AlterTableIT |
|   | phoenix.end2end.RowTimestampIT |
|   | phoenix.end2end.index.IndexRebuildIncrementDisableCountIT |
|   | phoenix.end2end.index.PhoenixMRJobSubmitterIT |
|   | phoenix.end2end.PowerFunctionEnd2EndIT |
|   | phoenix.end2end.CostBasedDecisionIT |
|   | phoenix.end2end.CbrtFunctionEnd2EndIT |
|   | phoenix.end2end.NullIT |
|   | phoenix.end2end.PartialResultServerConfigurationIT |
|   | phoenix.end2end.ArithmeticQueryIT |
|   | phoenix.end2end.SortOrderIT |
|   | phoenix.end2end.UpperLowerFunctionIT |
|   | phoenix.end2end.AggregateQueryIT |
|   | phoenix.end2end.DropSchemaIT |
|   | phoenix.end2end.DefaultPhoenixMultiViewListProviderIT |
|   | phoenix.end2end.MD5FunctionIT |
|   | phoenix.end2end.IndexToolForNonTxGlobalIndexIT |
|   | phoenix.end2end.AppendOnlySchemaIT |
|   | phoenix.end2end.join.HashJoinGlobalIndexIT |
|   | phoenix.end2end.index.IndexUsageIT |
|   | phoenix.end2end.DerivedTableIT |
|   | phoenix.tx.TxCheckpointIT |
|   | phoenix.end2end.MetaDataEndPointIT |
|   | phoenix.end2end.NthValueFunctionIT |
|   | phoenix.rpc.PhoenixServerRpcIT |
|   | phoenix.end2end.LogicalTableNameIT |
|   | phoenix.end2end.index.txn.TxWriteFailureIT |
|   | phoenix.end2end.InListIT |
|   | phoenix.end2end.index.LocalImmutableTxIndexIT |
|   | phoenix.end2end.FirstValuesFunctionIT |
|   | phoenix.end2end.index.ImmutableIndexIT |
|   | phoenix.end2end.ModulusExpressionIT |
|   | phoenix.end2end.join.HashJoinNoIndexIT |
|   | phoenix.end2end.DropTableWithViewsIT |
|   | phoenix.end2end.KeyOnlyIT |
|   | phoenix.end2end.ConcurrentMutationsIT |
|   | phoenix.end2end.DynamicUpsertIT |
|   | phoenix.end2end.ReverseScanIT |
|   | phoenix.iterate.ChunkedResultIteratorIT |
|   | phoenix.end2end.index.ShortViewIndexIdIT |
|   | phoenix.end2end.index.MutableIndexFailureWithNamespaceIT |
|   | phoenix.end2end.IndexScrutinyToolIT |
|   | phoenix.end2end.ExpFunctionEnd2EndIT |
|   | phoenix.end2end.join.HashJoinCacheIT |
|   | phoenix.end2end.ConcurrentMutationsExtendedIT |
|   | phoenix.end2end.index.IndexMaintenanceIT |
|   | phoenix.end2end.UpgradeNamespaceIT |
|   | hadoop.hbase.regionserver.wal.WALRecoveryRegionPostOpenIT |
|   | phoenix.end2end.index.InvalidIndexStateClientSideIT |
|   | phoenix.schema.stats.NamespaceEnabledStatsCollectorIT |
|   | phoenix.end2end.PhoenixDriverIT |
|   | phoenix.end2end.SCNIT |
|   | phoenix.end2end.join.SortMergeJoinGlobalIndexIT |
|   | phoenix.end2end.join.SortMergeJoinNoSpoolingIT |
|   | phoenix.end2end.index.AsyncIndexDisabledIT |
|   | phoenix.end2end.ToDateFunctionIT |
|   | phoenix.end2end.SumFunctionIT |
|   | phoenix.end2end.ArraysWithNullsIT |
|   | phoenix.tx.TransactionIT |
|   | phoenix.end2end.CreateTableIT |
|   | phoenix.end2end.QueryExecWithoutSCNIT |
|   | phoenix.end2end.DecodeFunctionIT |
|   | phoenix.end2end.SqrtFunctionEnd2EndIT |
|   | phoenix.end2end.join.SubqueryUsingSortMergeJoinIT |
|   | phoenix.end2end.index.GlobalMutableTxIndexIT |
|   | phoenix.end2end.PrimitiveTypeIT |
|   | phoenix.end2end.EmptyColumnIT |
|   | phoenix.end2end.ViewIT |
|   | phoenix.end2end.ProjectArrayElemAfterHashJoinIT |
|   | phoenix.schema.tool.SchemaToolExtractionIT |
|   | phoenix.end2end.index.MutableIndexSplitReverseScanIT |
|   | phoenix.end2end.RoundFloorCeilFuncIT |
|   | phoenix.end2end.ReadOnlyIT |
|   | phoenix.end2end.DropIndexedColsIT |
|   | phoenix.end2end.join.HashJoinMoreIT |
|   | phoenix.end2end.DynamicColumnIT |
|   | phoenix.end2end.RowValueConstructorIT |
|   | phoenix.end2end.AlterSessionIT |
|   | phoenix.monitoring.IndexMetricsIT |
|   | phoenix.end2end.index.MutableIndexRebuilderIT |
|   | phoenix.end2end.MathPIFunctionEnd2EndIT |
|   | phoenix.end2end.MappingTableDataTypeIT |
|   | phoenix.end2end.Array1IT |
|   | phoenix.end2end.ExplainPlanWithStatsDisabledIT |
|   | phoenix.end2end.TenantSpecificViewIndexIT |
|   | phoenix.end2end.MergeViewIndexIdSequencesToolIT |
|   | phoenix.end2end.ArrayRemoveFunctionIT |
|   | phoenix.end2end.index.IndexMetadataIT |
|   | phoenix.end2end.TenantSpecificTablesDMLIT |
|   | phoenix.end2end.CaseStatementIT |
|   | phoenix.end2end.PermissionNSEnabledIT |
|   | phoenix.end2end.RegexpSubstrFunctionIT |
|   | phoenix.end2end.PropertiesInSyncIT |
|   | phoenix.end2end.index.txn.MutableRollbackIT |
|   | phoenix.end2end.GroupByIT |
|   | phoenix.end2end.TenantSpecificTablesDDLIT |
|   | phoenix.end2end.index.IndexVerificationOutputRepositoryIT |
|   | phoenix.end2end.OrderByWithSpillingIT |
|   | phoenix.end2end.OrphanViewToolIT |
|   | phoenix.end2end.RangeScanIT |
|   | phoenix.end2end.UnionAllIT |
|   | phoenix.end2end.PhoenixTTLToolIT |
|   | phoenix.end2end.PercentileIT |
|   | phoenix.end2end.UpsertValuesIT |
|   | phoenix.end2end.UpsertWithSCNIT |
|   | phoenix.monitoring.PhoenixMetricsDisabledIT |
|   | phoenix.end2end.BackwardCompatibilityIT |
|   | phoenix.end2end.PartialScannerResultsDisabledIT |
|   | phoenix.schema.stats.NamespaceDisabledStatsCollectorIT |
|   | phoenix.end2end.QueryLoggerIT |
|   | phoenix.jdbc.SecureUserConnectionsIT |
|   | phoenix.end2end.AlterMultiTenantTableWithViewsIT |
|   | phoenix.end2end.LnLogFunctionEnd2EndIT |
|   | phoenix.end2end.AlterTableWithViewsIT |
|   | phoenix.end2end.MaxLookbackIT |
|   | phoenix.end2end.CoalesceFunctionIT |
|   | phoenix.end2end.SkipScanQueryIT |
|   | phoenix.end2end.index.LocalImmutableNonTxIndexIT |
|   | phoenix.end2end.SpooledTmpFileDeleteIT |
|   | phoenix.end2end.IndexRepairRegionScannerIT |
|   | phoenix.end2end.index.GlobalIndexOptimizationIT |
|   | phoenix.end2end.SystemCatalogRegionObserverIT |
|   | phoenix.end2end.BackwardCompatibilityForSplittableSyscatIT |
|   | phoenix.end2end.MetaDataEndpointImplIT |
|   | phoenix.execute.PartialCommitIT |
|   | phoenix.end2end.ViewUtilIT |
|   | phoenix.trace.PhoenixTableMetricsWriterIT |
|   | phoenix.end2end.OctetLengthFunctionEnd2EndIT |
|   | phoenix.end2end.RegexpSplitFunctionIT |
|   | phoenix.schema.stats.NoOpStatsCollectorIT |
|   | phoenix.end2end.OnDuplicateKeyIT |
|   | phoenix.end2end.ViewIndexIdRetrieveIT |
|   | phoenix.end2end.FirstValueFunctionIT |
|   | phoenix.end2end.index.MutableIndexReplicationIT |
|   | phoenix.end2end.IndexScrutinyToolForTenantIT |
|   | phoenix.end2end.WALAnnotationIT |
|   | phoenix.end2end.AlterAddCascadeIndexIT |
|   | phoenix.end2end.index.GlobalIndexCheckerIT |


| Subsystem | Report/Notes |
|----------:|:-------------|
| Docker | ClientAPI=1.41 ServerAPI=1.41 base: https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1295/14/artifact/yetus-general-check/output/Dockerfile |
| GITHUB PR | https://github.com/apache/phoenix/pull/1295 |
| Optional Tests | dupname asflicense shellcheck shelldocs hadolint javac javadoc unit xml compile |
| uname | Linux b8a0afa9db73 4.15.0-153-generic #160-Ubuntu SMP Thu Jul 29 06:54:29 UTC 2021 x86_64 x86_64 x86_64 GNU/Linux |
| Build tool | maven |
| Personality | dev/phoenix-personality.sh |
| git revision | master / 19b22604e |
| Default Java | Ubuntu-11.0.9.1+1-Ubuntu-0ubuntu1.20.04 |
| javadoc | https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1295/14/artifact/yetus-general-check/output/branch-javadoc-root.txt |
| shellcheck | https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1295/14/artifact/yetus-general-check/output/diff-patch-shellcheck.txt |
| javadoc | https://ci-hadoop.apache.org/job/Phoenix/job/Phoenix-PreCommit-GitHub-PR/job/PR-1295/14/artifact/yetus-general-check/output/patch-javadoc-root.txt |
…